refactor: confirm delete modal changes
