+++ src/sections/file/file-action-buttons/edit-file-dropdown/delete-file-button/confirm-delete-file-modal/ConfirmDeleteFileModal.module.scss
