+++ src/stories/file/file-action-buttons/edit-file-menu/ConfirmDeleteFileModal.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta, StoryObj } from '@storybook/react'
 import { WithI18next } from '../../../WithI18next'
 import { WithSettings } from '../../../WithSettings'
-import { ConfirmDeleteFileModal } from '@/sections/file/file-action-buttons/edit-file-dropdown/delete-file-button/ConfirmDeleteFileModal'
+import { ConfirmDeleteFileModal } from '@/sections/file/file-action-buttons/edit-file-dropdown/delete-file-button/confirm-delete-file-modal/ConfirmDeleteFileModal'
 
 const meta: Meta<typeof ConfirmDeleteFileModal> = {
   title: 'Sections/File Page/Action Buttons/EditFileDropdown/ConfirmDeleteFileModal',
